Fix incorrect type used to track transient resource last use time
Truncated frame start time to 32 bits, meant resources would be freed and
recreated every frame.

+++ Engine/Runtime/Render/RenderManager.h
@@ -77,7 +77,7 @@ private:
          * and also whether the resource is available for reuse in the current
          * frame.
          */
-        uint32_t                lastUsedFrameStartTime;
+        uint64_t                lastUsedFrameStartTime;
     };
 
     struct TransientBuffer : TransientResource
